Fix .jar file error with Gradle
