+++ build.gradle
@@ -48,7 +48,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Duke"
+    mainClassName = "betago.Main"
 }
 
 shadowJar {
