+++ src/main/java/betago/Launcher.java
@@ -1,12 +1,39 @@
 package betago;
 
+import java.io.IOException;
+
+import betago.controller.MainWindow;
 import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
 
 /**
  * A launcher class to workaround classpath issues.
  */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
+public class Launcher extends Application {
+    private Duke duke = new Duke();
+
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            stage.setTitle("BetaGo");
+            fxmlLoader.<MainWindow>getController().setDuke(duke);
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Launches the application.
+     */
+    public void launch() {
+        Application.launch();
     }
 }
