+++ src/main/java/betago/Main.java
@@ -1,35 +1,15 @@
 package betago;
 
-
-import java.io.IOException;
-
-import betago.controller.MainWindow;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
 /**
  * A GUI for Duke using FXML.
  */
-public class Main extends Application {
-
-    private Duke duke = new Duke();
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            stage.setTitle("BetaGo");
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+public class Main {
+    /**
+     * Calls the launch method in launcher.
+     */
+    public static void main(String[] args) {
+        Launcher launcher = new Launcher();
+        launcher.launch();
     }
 }
